split gregor
