+++ shared/constants/gregor.js
@@ -1,20 +1,8 @@
 // @flow
-import * as RPCTypesGregor from '../constants/types/flow-types-gregor'
-import * as RPCTypes from '../constants/types/flow-types'
+import * as Types from './types/gregor'
+import * as RPCTypes from './types/flow-types'
 
-export type NonNullGregorItem = {
-  md: RPCTypesGregor.Metadata,
-  item: RPCTypesGregor.Item,
-}
-
-export type MsgMap = {[key: string]: ?NonNullGregorItem}
-
-export type State = {
-  reachability: RPCTypes.Reachability,
-  seenMsgs: MsgMap,
-}
-
-const initialState: State = {
+const initialState: Types.State = {
   reachability: {reachable: RPCTypes.reachabilityReachable.unknown},
   seenMsgs: {},
 }
