+++ shared/constants/reducer.js
@@ -7,7 +7,7 @@ import type {State as Devices} from '../constants/types/devices'
 import type {State as Entity} from '../constants/types/entities'
 import type {State as Engine} from '../constants/types/engine'
 import type {State as Favorite} from '../constants/types/favorite'
-import type {State as Gregor} from '../constants/gregor'
+import type {State as Gregor} from '../constants/types/gregor'
 import type {State as Login} from '../constants/login'
 import type {State as Notification} from '../constants/notifications'
 import type {State as Pgp} from '../constants/pgp'
