+++ shared/reducers/gregor.js
@@ -1,18 +1,16 @@
 // @flow
+import * as Types from '../constants/types/gregor'
 import * as Constants from '../constants/gregor'
 import * as GregorGen from '../actions/gregor-gen'
 import keyBy from 'lodash/keyBy'
 
-export default function(
-  state: Constants.State = Constants.initialState,
-  action: GregorGen.Actions
-): Constants.State {
+export default function(state: Types.State = Constants.initialState, action: GregorGen.Actions): Types.State {
   switch (action.type) {
     case GregorGen.resetStore:
       return {...Constants.initialState}
     case GregorGen.updateSeenMsgs:
       // TODO do we ever use this?
-      const newMsgs: Constants.MsgMap = keyBy(action.payload.seenMsgs, m => m.md.msgID.toString('base64'))
+      const newMsgs: Types.MsgMap = keyBy(action.payload.seenMsgs, m => m.md.msgID.toString('base64'))
       return {
         ...state,
         seenMsgs: {
